refactor(secrets): split to foss and pro
