+++ .gitignore
@@ -38,3 +38,4 @@ __debug_bin*
 /events.log
 /tasks.log
 /task_results.log
+/pro/
