+++ services/server_services/access_key_serializer.go
@@ -4,7 +4,7 @@ import (
 	"github.com/semaphoreui/semaphore/db"
 )
 
-type AccessKeyKeyDeserializer interface {
+type AccessKeyDeserializer interface {
 	DeserializeSecret(key *db.AccessKey) (string, error)
 	SerializeSecret(key *db.AccessKey) error
 	DeleteSecret(key *db.AccessKey) error
