+++ services/server_services/access_key_serializer_vault_foss.go
@@ -1,0 +1,30 @@
+//go:build !pro
+
+package server_services
+
+import (
+	"github.com/semaphoreui/semaphore/db"
+)
+
+type VaultAccessKeyDeserializer struct {
+}
+
+func NewVaultAccessKeyDeserializer(
+	accessKeyRepo db.AccessKeyManager,
+	secretStorageRepo db.SecretStorageRepository,
+	encryptionService AccessKeyEncryptionService,
+) *VaultAccessKeyDeserializer {
+	return &VaultAccessKeyDeserializer{}
+}
+
+func (d *VaultAccessKeyDeserializer) DeleteSecret(key *db.AccessKey) error {
+	return nil
+}
+
+func (d *VaultAccessKeyDeserializer) SerializeSecret(key *db.AccessKey) (err error) {
+	return nil
+}
+
+func (d *VaultAccessKeyDeserializer) DeserializeSecret(key *db.AccessKey) (res string, err error) {
+	return
+}
